+++ packages/vue/types/Link.d.ts
@@ -1,7 +1,7 @@
 import Vue, { ComponentOptions } from "vue";
 import { SessionLocation } from "@hickory/root";
 export interface LinkComponent extends Vue {
-    to: string;
+    name: string;
     params?: object;
     hash?: string;
     query?: any;