+++ packages/vue/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 2.0.0-alpha.3
 
 * Rename `<curi-link>`'s `to` prop to `name`.
 * Split `<curi-link>` into two components. `<curi-link>` is static (regular slot) while `<curi-async-link>` tracks navigating state (scoped slot).